Create ChangeFieldName.pyt
Uses Alter Field to change the field name. Can select one field at a time or enter a string that will be deleted from all fields in the feature if that string is in the field name. This is specifically to alter fields that the ArcGIS geocoding services creates from your table of addresses. That service adds "USER_{your field name}" to most of your fields and this can create data problems if all you're trying to do is to create a point feature class from a table.